# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/parcelas": {
      "post": {
        "operationId": "ParcelasController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateParcelaDto"
              },
              "examples": {
                "default": {
                  "value": {
                    "ensayoId": 1,
                    "bloqueId": 1,
                    "tratamientoId": 1,
                    "posXGrid": 2,
                    "posYGrid": 3
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Parcela creada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Parcela"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crear una parcela",
        "tags": [
          "parcelas"
        ]
      },
      "get": {
        "operationId": "ParcelasController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Parcela"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar parcelas",
        "tags": [
          "parcelas"
        ]
      }
    },
    "/api/v1/parcelas/{id}": {
      "get": {
        "operationId": "ParcelasController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Parcela"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtener parcela por ID",
        "tags": [
          "parcelas"
        ]
      },
      "patch": {
        "operationId": "ParcelasController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateParcelaDto"
              },
              "examples": {
                "default": {
                  "value": {
                    "posXGrid": 5,
                    "posYGrid": 6
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Parcela"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Actualizar parcela",
        "tags": [
          "parcelas"
        ]
      },
      "delete": {
        "operationId": "ParcelasController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "deleted": true
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Eliminar parcela",
        "tags": [
          "parcelas"
        ]
      }
    },
...
    "/api/v1/momentos": {
      "post": {
        "operationId": "MomentosController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMomentoDto"
              },
              "examples": {
                "default": {
                  "value": {
                    "aplicacionId": 1,
                    "nombreMomento": "7DDA",
                    "diasDespuesAplicacion": 7,
                    "fechaEvaluacion": "2025-04-01"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Momento creado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MomentoEvaluacion"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crear un momento de evaluación",
        "tags": [
          "momentos"
        ]
      },
      "get": {
        "operationId": "MomentosController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MomentoEvaluacion"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar momentos de evaluación",
        "tags": [
          "momentos"
        ]
      }
    },
    "/api/v1/momentos/{id}": {
      "get": {
        "operationId": "MomentosController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MomentoEvaluacion"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtener momento de evaluación por ID",
        "tags": [
          "momentos"
        ]
      },
      "patch": {
        "operationId": "MomentosController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMomentoDto"
              },
              "examples": {
                "default": {
                  "value": {
                    "nombreMomento": "14DDA",
                    "diasDespuesAplicacion": 14
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MomentoEvaluacion"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Actualizar momento de evaluación",
        "tags": [
          "momentos"
        ]
      },
      "delete": {
        "operationId": "MomentosController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "deleted": true
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Eliminar momento de evaluación",
        "tags": [
          "momentos"
        ]
      }
    },
...
  },
  "info": {
    "title": "API de Ensayos Agronómicos",
    "description": "Documentación de la API para el sistema de gestión de ensayos",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateParcelaDto": {
        "type": "object",
        "properties": {
          "ensayoId": {
            "type": "number",
            "example": 1,
            "description": "ID del ensayo"
          },
          "bloqueId": {
            "type": "number",
            "example": 1,
            "description": "ID del bloque"
          },
          "tratamientoId": {
            "type": "number",
            "example": 1,
            "description": "ID del tratamiento"
          },
          "posXGrid": {
            "type": "object",
            "example": 1
          },
          "posYGrid": {
            "type": "object",
            "example": 2
          }
        },
        "required": [
          "ensayoId",
          "bloqueId",
          "tratamientoId"
        ]
      },
      "Parcela": {
        "type": "object",
        "properties": {}
      },
      "UpdateParcelaDto": {
        "type": "object",
        "properties": {
          "ensayoId": {
            "type": "number",
            "example": 1,
            "description": "ID del ensayo"
          },
          "bloqueId": {
            "type": "number",
            "example": 1,
            "description": "ID del bloque"
          },
          "tratamientoId": {
            "type": "number",
            "example": 1,
            "description": "ID del tratamiento"
          },
          "posXGrid": {
            "type": "object",
            "example": 1
          },
          "posYGrid": {
            "type": "object",
            "example": 2
          }
        }
      },
...
      "CreateMomentoDto": {
        "type": "object",
        "properties": {
          "aplicacionId": {
            "type": "number",
            "example": 1,
            "description": "ID de la aplicación a la que pertenece"
          },
          "nombreMomento": {
            "type": "string",
            "example": "7DDA",
            "maxLength": 50
          },
          "diasDespuesAplicacion": {
            "type": "object",
            "example": 7
          },
          "fechaEvaluacion": {
            "type": "object",
            "example": "2025-04-01"
          }
        },
        "required": [
          "aplicacionId",
          "nombreMomento"
        ]
      },
      "MomentoEvaluacion": {
        "type": "object",
        "properties": {}
      },
      "UpdateMomentoDto": {
        "type": "object",
        "properties": {
          "aplicacionId": {
            "type": "number",
            "example": 1,
            "description": "ID de la aplicación a la que pertenece"
          },
          "nombreMomento": {
            "type": "string",
            "example": "7DDA",
            "maxLength": 50
          },
          "diasDespuesAplicacion": {
            "type": "object",
            "example": 7
          },
          "fechaEvaluacion": {
            "type": "object",
            "example": "2025-04-01"
          }
        }
      },
...
    }
  }
}